# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Kit HTTP API
  version: 1.0.0
  description: >-
    Authenticated M001 command/query API. Clients must ignore unknown response
    fields. Identifiers and cursors are opaque and must not be parsed. Executor
    mutation outcomes are retained for the declared retry window. Executor cursor
    recovery snapshots contain authorized processes, terminals, and current
    attachments, including replacement history viewers on exited or interrupted
    terminals; detached, stale, and restart-invalidated attachments are omitted.
    Native PTY primitives do not imply that a production attempt-owned executor
    profile can expose a PTY.
...
servers:
  - url: http://localhost
security:
  - bearerAuth: []
...
paths:
  /v1/projects:
    post:
      operationId: createProject
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      responses:
        '201':
          $ref: '#/components/responses/ResourceCreated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getProject
      tags: [Projects]
      responses:
        '200':
          description: Project projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/projects/{project_id}/threads:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listThreads
      tags: [Threads]
      responses:
        '200':
          description: Thread projections.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Thread'
                additionalProperties: true
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
    post:
      operationId: createThread
      tags: [Threads]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThread'
      responses:
        '201':
          $ref: '#/components/responses/ResourceCreated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/threads/{thread_id}:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getThread
      tags: [Threads]
      responses:
        '200':
          description: Thread projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/threads/{thread_id}/archive:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: setThreadArchived
      tags: [Threads]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetThreadArchived'
      responses:
        '200':
          $ref: '#/components/responses/ResourceUpdated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/threads/{thread_id}/deletion:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: initiateThreadDeletion
      tags: [Threads]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedVersion'
      responses:
        '202':
          $ref: '#/components/responses/ResourceAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '423':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/deletion-jobs/{deletion_job_id}:
    parameters:
      - $ref: '#/components/parameters/DeletionJobId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getDeletionJob
      tags: [Retention]
      responses:
        '200':
          description: Current auditable deletion-job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionJob'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/threads/{thread_id}/runs:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: startRun
      tags: [Runs]
      description: >-
        Durably stores the message artifact, then atomically records the input reference and queued
        run. Reusing Idempotency-Key with the same request replays the receipt; different input
        conflicts. A 202 response does not indicate run completion.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRun'
      responses:
        '202':
          $ref: '#/components/responses/ResourceAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/projects/{project_id}/runs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listRuns
      tags: [Runs]
      responses:
        '200':
          description: Run projections.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Run'
                additionalProperties: true
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRun
      tags: [Runs]
      responses:
        '200':
          description: Run projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: cancelRun
      tags: [Runs]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedVersion'
      responses:
        '202':
          $ref: '#/components/responses/ResourceAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/cost:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRunCost
      tags: [Runs]
      responses:
        '200':
          description: Durable reconciled run usage and cost, with null values while unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCost'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/prompts:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRunPrompts
      tags: [Runs]
      responses:
        '200':
          description: Canonical prompt digests, stable-prefix boundary, and budget measurement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPrompts'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/transcript:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRunTranscript
      tags: [Runs]
      responses:
        '200':
          description: Public, reasoning-free semantic model and tool transcript.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTranscript'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/events:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/LastEventId'
    get:
      operationId: getRunEvents
      tags: [Events]
      description: >-
        Returns a page by default; with `Accept: text/event-stream`, follows only
        this run and resumes from `Last-Event-ID`.
      responses:
        '200':
          description: Durable event page or live run-scoped SSE stream.
          headers:
            X-Kit-Cursor:
              schema:
                $ref: '#/components/schemas/OpaqueStreamCursor'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/input:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: provideRunInput
      tags: [Runs]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvideRunInput'
      responses:
        '200':
          $ref: '#/components/responses/ResourceUpdated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '415':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
  /v1/threads/{thread_id}/events:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/LastEventId'
    get:
      operationId: getThreadEvents
      tags: [Events]
      description: >-
        Returns a page by default; with `Accept: text/event-stream`, follows only
        this thread and its runs and resumes from `Last-Event-ID`.
      responses:
        '200':
          description: Durable event page or live thread-scoped SSE stream.
          headers:
            X-Kit-Cursor:
              schema:
                $ref: '#/components/schemas/OpaqueStreamCursor'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '504':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Transport-issued bearer credential. Remote deployments validate OIDC or mTLS before routing.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Retained with the mutation receipt for durable retry and conflict detection.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        pattern: '^[!-~]+$'
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
...
  responses:
...
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthenticated:
      description: Authentication failed before resource dispatch.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ResourceCreated:
      description: Resource created synchronously.
      headers:
        Location:
          required: true
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceReceipt'
...
  schemas:
...
    CreateProject:
      type: object
      required: [id]
      properties:
        id:
          $ref: '#/components/schemas/ProjectId'
      additionalProperties: true
...